Fix ux bugs, add play screen quit popup, refactor.
- Improve / fix user interaction in the main menu:

        - When user presses ESC on main menu, unfocus & disable
          interaction for everything behind the quit popup.

        - Restore focus & interaction when the quit popup is cancelled
          by pressing ESC again or clicking CANCEL button.

        - Add fade in/out effect to the quit popup.

        - When ENTER is pressed during the quit popup or the QUIT button
          is clicked, the popup will completely finish its fade out
          effect AND be removed from the screen BEFORE switching screens.

        - Reset the entire main menu screen state upon clicking a menu
          item and going to a different screen.

        - Force a mouse movement event every time the screen is shown,
          to update any mouse-dependent effects that were left over from
          last time the screen was shown.

- Implement classic play screen quit popup, and improve / fix user
  interaction. When user presses ESC on classic play screen, the
  following will occur:

        - Unfocus & disable interaction for everything behind the popup.

        - Restore focus & interaction when the popup is cancelled by
          pressing ESC again or clicking CANCEL button.

        - Pressing ENTER or clicking QUIT button quits the play screen
          and goes back to the previous screen (or the main menu,
          assuming for some reason that the initial screen was the play screen.)

        - Quit popup has a fade in/out effect when appearing / disappearing.

        - When ENTER is pressed during the quit popup or the QUIT button
          is clicked, the popup will completely finish its fade out
          effect AND be removed from the screen BEFORE switching screens.

        - Reset the entire play screen state upon quitting / submitting
          the popup.

        - Force a mouse movement event every time the screen is shown,
          to update any mouse-dependent effects that were left over from
          last time the screen was shown.

        - Pressing ESC while having keyboard focus inside of the chatbox
          textfield still brings up the quit popup, and if ESC is
          pressed again to cancel the quit popup, then the keyboard
          focus will be restored to the textfield at the same location
          it was at.

- Debugging input shortcut changes: LEFT key on classic play screen no
  longer changes screens, and ESC no longer quits the entire application
  on the same screen. RIGHT key still goes to the peril play screen,
  however. Z disables play map ux, X re-enables it.

- Scene2d input hit boxes are off vertically by about 3 pixels; this
  could be an LWJGL bug (which LibGDX / Scene2d delegates to) on OS X,
  but fixing it requires shifting the OpenGL viewport origin to 0, -3 via
  the resize callback in every Screen. So far this mainly affects buttons.

Refactoring:

- Add methods to reset, enable, & disable PlayMapActor. Reset resets the
  play map state to default - all countries have UNOWNED
  CountryImageState, and all country army text is reset to a count of 0.
  Disable hides the territory hover text, ends any hover or click effects
  on the play map, and prevents further user interaction. Enable shows the
  territory hover text, re-enables user interaction, and uses an updated
  mouse position to set any new hover effects.

- Extract QuitPopup from MainMenuScreen into its own class so that it
  can be re-used on other screens (such as the ClassicPlayScreen).

- Instead of accessing GraphicsSettings.REFERENCE_WIDTH/HEIGHT/SIZE
  statically, computing screen scaling manually, and globally /
  statically accessing Gdx.graphics.getWidth/Height everywhere, create
  ScreenSize that does all of this automatically, and inject it into
  classes that need this functionality. This solves the problem of classes
  having to manually deal with getting / updating the screen size
  themselves, which resulted in code duplication. Now all of this
  functionality is focused into one class, and it's more obvious which
  classes are using this functionality, as they now have a dependency on
  ScreenSize. Less code, & less potential for bugs.

- Create LibGdxScreenSize implementation of ScreenSize interface, that
  gets passed a Graphics instance in its constructor (way more OOP than
  calling Gdx.graphics everywhere.)

  Music:

        - Get rid of ScreenMusic implementations for each Screen, as
          they are all the same. Eliminate duplication by creating a
          DefaultScreenMusic implementation, and pass this to each
          Screen.

        - DefaultScreenMusic won't #start or #stop music if it is not
          enabled in MusicSettings. This simplifies callers of #start &
          #stop because they don't have to check if the music is enabled first.

        - Make MusicSettings instantiable with getters, and inject it
          into the DefaultScreenMusic constructor, to achieve a more
          object-oriented design (instead of referencing MusicSettings statically,
          i.e., de-globalizing it). This way, each screen could also
          have different music settings.

- Do not create ClassicPlayScreen dialogs/popups in
  PlayScreenWidgetFactory any more because they require the implementation
  of callbacks that only ClassicPlayScreen has the knowledge to
  complete, and is too much for a factory to deal with. They are
  instantiated directly in ClassicPlayScreen now, in order to be able to
  take action in onSubmit, onShow, & onHide.

- Create PlayMapActor via PlayScreenWidgetFactory, instead of passing it
  in the ClassicPlayScreen constructor. This simplifies the
  ClassicPlayScreen constructor and makes PlayScreenWidgetFactory more
  useful (it's purpose is to create many widgets for the
  ClassicPlayScreen, so they don't have to be passed in directly via the
  constructor. As more widgets are added to the screen, the constructor
  does not need to add any more parameters).

- Make GraphicsSettings.INITIAL_WINDOW_WIDTH,
  GraphicsSettings.INITIAL_WINDOW_HEIGHT,
  GraphicsSettings.REFERENCE_SCREEN_WIDTH, &
  GraphicsSettings.REFERENCE_SCREEN_HEIGHT integers instead of floats.  It
  simplifies their use in other places and they will always be integers.

- Provide MouseInput#position method that re-uses a Vector2 instance
  because it is needed by PlayMapActor & HoveredTerritoryTextActor, and
  simplifies their code, since they don't have to place the individual x
  & y values into a Vector2.

- Rename MouseInput#getHoverX & MouseInput#getHoverY to MouseInput#x &
  MouseInput#y, respectively, because it's the position of the mouse
  regardless of whether or not the mouse is hovering or clicking or
  whatever, and because it makes for more succinct code for callers of
  these methods.

- Rename AbstractOkPopup & AbstractOkCancelPopup to OkPopup &
  CancelPopup, respectively, because although they are technically
  abstract classes, that word should only be added to classes that are not
  intended to be used without extending them; however these classes can be
  used as-is (without extending), they just require implementing
  a few callbacks.

- Rearrange constructor parameters in play screens and play screen
  factories for better consistency.

- Remove Arguments.checkIsNotNull calls from private methods in
  PlayMapActor (they were previously refactored from public to private).

- Remove commented out debugging code from MainMenuScreen &
  PerilPlayScreen.

- Add missing // @formatter:on tag at the end of the
  PlayMapInputDetectionFactory constructor.

- Update code style on some classes where needed.

+++ client/src/main/java/com/forerunnergames/peril/client/settings/GraphicsSettings.java
@@ -9,10 +9,10 @@ import com.forerunnergames.tools.common.Classes;
 public final class GraphicsSettings
 {
   // @formatter:off
-  public static final float                 MIN_INITIAL_WINDOW_WIDTH          = 640;
-  public static final float                 MIN_INITIAL_WINDOW_HEIGHT         = 480;
-  public static final float                 REFERENCE_SCREEN_WIDTH            = 1920;
-  public static final float                 REFERENCE_SCREEN_HEIGHT           = 1080;
+  public static final int                   MIN_INITIAL_WINDOW_WIDTH          = 640;
+  public static final int                   MIN_INITIAL_WINDOW_HEIGHT         = 480;
+  public static final int                   REFERENCE_SCREEN_WIDTH            = 1920;
+  public static final int                   REFERENCE_SCREEN_HEIGHT           = 1080;
   public static final Vector2               REFERENCE_SCREEN_SIZE             = new Vector2 (REFERENCE_SCREEN_WIDTH, REFERENCE_SCREEN_HEIGHT);
   public static final boolean               TEXTURE_MIPMAPPING                = true;
   public static final Texture.TextureFilter TEXTURE_MINIFICATION_FILTER       = Texture.TextureFilter.MipMapLinearNearest;
